+++ src/Kernel/src/SimTypes.h
@@ -29,6 +29,20 @@
 
 namespace SimTypes {
 
+    enum class LightState {
+        Navigation = 0x0001,
+        Beacon = 0x0002,
+        Landing = 0x0004,
+        Taxi = 0x0008,
+        Strobe = 0x0010,
+        Panel = 0x0020,
+        Recognition = 0x0040,
+        Wing = 0x0080,
+        Logo = 0x0100,
+        Cabin = 0x0200
+    };
+    Q_DECLARE_FLAGS(LightStates, SimTypes::LightState)
+
     enum class SurfaceType {
         Unknown,
         Concrete,
@@ -213,5 +227,7 @@ namespace SimTypes {
         }
     }
 }
+
+Q_DECLARE_OPERATORS_FOR_FLAGS(SimTypes::LightStates)
 
 #endif // SIMTYPES_H